Allow multiple accounts to use same inbound group session
Consider the following scenario: Alice uses Element to create a room, invite
Bob and Bob's bot (which share a Pantalaimon instance), and send a message.
Both Bob and Bob's bot can decrypt the message, but only until Pantalaimon
restarts.

Before this change, Pantalaimon would create an inbound group session, store it
in memory (inbound_group_store), and (try to) persist a corresponding record in
the megolminboundsessions table, one for Bob and one for Bob's bot. This worked
correctly in memory, but only one of the records would be persisted because
megolminboundsessions prevented inserting multiple records with the same
session_id even if the account_id was different.

This change modifies the schema to support multiple megolminboundsessions
records having the same session_id but different account_id.

A separate commit will have migration logic similar to the following:
https://gist.github.com/codemonium/14bc345313f93e08a44eb13720937958