Allow multiple accounts to use same Olm session
Suppose that Bob and Bob's bot share a Pantalaimon instance. When Bob
invites his bot to a room and sends a message, Pantalaimon tries to
create two olmsessions records: one for Bob and one for his bot. Both
records have the same session_id, which would prevent the second record
from being persisted. I cannot guarantee that this prevents the user
from decrypting the initial message, but it has happened before in this
context.

This change modifies the schema to support multiple olmsessions records
having the same session_id but different account_id.