Persist all Megolm sessions (for room) on join
When a user previews an encrypted public room, we store all its Megolm
sessions in memory but not in the database. Before this change, if the
user joined the room and Pantalaimon was restarted, the user would no
longer be able to decrypt the existing messages. Now, we save the
sessions to the database when the user joins the room.