Prevent BAD_MESSAGE_MAC errors
When using pantalaimon, there is a chance that both pantalaimon and
matrix-nio try to handle the to-device events from the same sync
response, which causes BAD_MESSAGE_MAC errors. This change adds a method
that uses the next_batch token to decide whether the to-device events
have already been handled.